requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.12
idna==3.3
lxml==5.3.2
orjson==3.10.16
python-dotenv==1.1.0
pytz==2025.2
requests==2.26.0
ruff==0.11.5
urllib3==1.26.5
